changed structure of move limits (one type of move limit per file)
