+++ sao/move_limits/__init__.py
@@ -1,2 +1,4 @@
-from .move_limit import Bounds, MoveLimit, AdaptiveMoveLimit
+from .move_limit import MoveLimit, MoveLimitFraction
+from .bounds import Bounds
+from .adaptive_move_limits import AdaptiveMoveLimit
 from .mixed_move_limit import MixedMoveLimit
